pyproject.toml
[tool.poetry]
name = "ai-service"
version = "0.1.0"
description = ""
authors = ["thwolter <thwolter@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.2"}
uvicorn = "^0.31.1"
langchain = "^0.3.25"
langchain-openai = "^0.3.21"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.9.1"
pytest = "^8.3.3"
yake = {git = "https://github.com/LIAAD/yake"}
fastapi-cache2 = {extras = ["redis"], version = "^0.2.2"}
diskcache = "^5.6.3"
aioredis = "^2.0.1"
langchain-community = "^0.3.24"
pyjwt = "^2.10.1"
sentry-sdk = {extras = ["fastapi"], version = "^2.19.2"}
numpy = "^2.3"
openmeter = "^1.0.0b186"
riskgpt = {git = "https://github.com/thwolter/riskgpt.git", rev = "develop"}
pytest-asyncio = "^1.0.0"
aiobreaker = "^1.2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-dotenv = "^0.5.2"
isort = "^5.13.2"
ruff = "^0.6.9"
mypy = "^1.9.0"
black = "^24.3.0"
pre-commit = "^3.6.0"
mkdocs = "^1.6.1"
mkdocs-material = "^9.6.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.format]
quote-style = "single"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100
py_version = "312"

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "webtest: marks tests that access external APIs (select with '-m webtest')",
    "tryfirst: marks tests that should be run first",
    "no_quota_patch: do not patch quota service in this test",
]
dotenv_file = ".env.test"
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
